+++ go/lib.go
@@ -109,7 +109,7 @@ func stopLocked() {
 }
 
 // Start initializes the Tailscale node.
-func Start(hostname, authKey, controlURL, stateDir string) (err error) {
+func Start(hostname, authKey, controlURL, stateDir string, ephemeral bool) (err error) {
 	mu.Lock()
 	defer mu.Unlock()
 
@@ -157,6 +157,7 @@ func Start(hostname, authKey, controlURL, stateDir string) (err error) {
 		ControlURL: controlURL,
 		Dir:        stateDir,
 		Store:      newStore,
+		Ephemeral:  ephemeral,
 		Logf: func(format string, args ...any) {
 			if atomic.LoadInt32(&LogLevel) >= 2 {
 				log.Printf("TSNET: "+format, args...)