+++ lib/src/worker/worker.dart
@@ -188,6 +188,7 @@ final class Worker {
   Future<void> start({
     required String hostname,
     required String authKey,
+    required bool ephemeral,
     required String controlUrl,
     required String stateDir,
   }) async {
@@ -196,6 +197,7 @@ final class Worker {
       _WorkerStartCommand(
         hostname: hostname,
         authKey: authKey,
+        ephemeral: ephemeral,
         controlUrl: controlUrl,
         stateDir: stateDir,
         hostNetworkSnapshot: hostNetworkSnapshot,